Break out Skill helpers from quests
