+++ src/main/java/com/questhelper/skills/agility/Agility.java
@@ -22,7 +22,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.questhelper.quests.agility;
+package com.questhelper.skills.agility;
 
 import com.questhelper.ItemCollections;
 import com.questhelper.QuestDescriptor;
